% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_labels}
\alias{add_labels}
\title{Add many labels to a dygraph}
\usage{
add_labels(x, events, ...)
}
\arguments{
\item{x}{a dygraph object (usually piped in)}

\item{events}{a vector x-axis markers}

\item{...}{additional arguments to dyEvent}
}
\description{
Add many labels to a dygraph
}
