% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{add_shades}
\alias{add_shades}
\title{Add many shaing areas to a dygraph}
\usage{
add_shades(x, periods, ...)
}
\arguments{
\item{x}{A dygraph object (usually piped in)}

\item{periods}{A df with start and stop pairs to mark the shaded areas}

\item{...}{additional arguments to dyEvent}
}
\description{
Add many shaing areas to a dygraph
}
