% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update-to-options.R
\name{add_time_column}
\alias{add_time_column}
\title{Add a time column to all trap-data.csv}
\usage{
add_time_column(project_path, hz = 5000)
}
\arguments{
\item{project_path}{project folder path}

\item{hz}{}
}
\description{
Add a time column to all trap-data.csv
}
