% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{allow_if}
\alias{allow_if}
\title{Allow reactivity to continue in shiny if expression is TRUE}
\usage{
allow_if(.if, is_shiny = T, ...)
}
\arguments{
\item{.if}{An expression that returns TRUE/FALSE (to be passed to an internal if statement)}

\item{is_shiny}{A logical. Default is TRUE.}

\item{...}{Additional arguments passed to shinyShowNotification}
}
\value{
Nothing. Will either stop or allow shiny reactivity to continue
}
\description{
Allow reactivity to continue in shiny if expression is TRUE
}
\examples{
# This will stop reactivity and showNotification in UI
allow_if(FALSE, message = 'Oops!', type = 'error')

# This will allow reactivity to continue showing nothing in UI
allow_if(TRUE, message = 'Oops!', type = 'error')

# Practice at consolse with is_shiny = F

allow_if(TRUE, is_shiny = F)
}
