% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cloud-viewer.R
\name{archive_project_to_googledrive}
\alias{archive_project_to_googledrive}
\title{Archive a lasertrapr project to googledrive.}
\usage{
archive_project_to_googledrive(project, email)
}
\arguments{
\item{project}{a character string. a lasertrapr project folder.}

\item{email}{a character string. email account associated with your google drive.}
}
\value{
nothing
}
\description{
Originally created for use with the lasertrapr cloud viewer, but can be used to backup data to goolge drive as well. 
If local changes are made and you want to update the google drive archive, running this function again will trash the old project folder on googledrive
and recreate it from scratch. Data is placed in a folder called "lasertrapr-cloud-storage" on googledrive. If this folder does not exist it will be created.
}
