% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemble_average.R
\name{avg_ensembles}
\alias{avg_ensembles}
\title{Ensemble Average}
\usage{
avg_ensembles(f, is_shiny = TRUE)
}
\arguments{
\item{is_shiny}{logical. is function being used in shiny or not.}

\item{trap_selected_project}{full path to currently selected project}
}
\value{

}
\description{
Ensemble Average
}
