% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hm-cp-analysis.R
\name{changepoint_analysis}
\alias{changepoint_analysis}
\title{Runs changepoint analysis, estimates more accurate event measurements, and saves data for ensemble averaging.}
\usage{
changepoint_analysis(
  measured_hm_events,
  hz,
  conversion,
  mv2nm,
  conditions,
  front_cp_method,
  back_cp_method,
  cp_running_var_window,
  ws,
  displacement_type
)
}
\arguments{
\item{measured_hm_events}{the results from measure_hm_events()}

\item{hz}{sampling frequency in Hz (default is 5000)}

\item{conversion}{}
}
\value{

}
\description{
Runs changepoint analysis, estimates more accurate event measurements, and saves data for ensemble averaging.
}
