% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hm_cp_analysis.R
\name{changepoint_and_ensemble}
\alias{changepoint_and_ensemble}
\title{Runs changepoint analysis, estimates more accurate event measurements, and saves data for ensemble averaging.}
\usage{
changepoint_and_ensemble(measured_hm_events, hz, conversion)
}
\arguments{
\item{measured_hm_events}{the results from measure_hm_events()}

\item{hz}{sampling frequency in Hz (default is 5000)}

\item{conversion}{}
}
\value{

}
\description{
Runs changepoint analysis, estimates more accurate event measurements, and saves data for ensemble averaging.
}
