% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{correlations}
\alias{correlations}
\title{Title}
\usage{
correlations(event_files_filtered, plot_colors)
}
\arguments{
\item{event_files_filtered}{}

\item{plot_colors}{}
}
\value{

}
\description{
Title
}
