% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{defend_if}
\alias{defend_if}
\title{Stops reactivity in shiny app if expression is TRUE}
\usage{
defend_if(.if, is_shiny = T, ...)
}
\arguments{
\item{.if}{An expression that returns TRUE/FALSE (to be passed to an internal if statement)}

\item{...}{Additional arguments passed to shinyShowNotification}
}
\value{
Nothing. Will either stop or allow shiny reactivity to continue
}
\description{
Stops reactivity in shiny app if expression is TRUE
}
\examples{
# This will stop reactivity and showNotification in UI
defend_if(TRUE, message = 'Oops!', type = 'error')

# This will allow reactivity to continue showing nothing in UI
defend_if(TRUE, message = 'Oops!', type = 'error')

# Practice at consolse with is_shiny = F

defend_if(TRUE, is_shiny = F)
}
