% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ensemble_average.R
\name{drc_forward_ee_2exp}
\alias{drc_forward_ee_2exp}
\title{xDP-release, xTP-binding self-starter (fct) for drm}
\usage{
drc_forward_ee_2exp()
}
\value{

}
\description{
xDP-release, xTP-binding self-starter (fct) for drm
}
