% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cloud-viewer.R
\name{drive_read_trap}
\alias{drive_read_trap}
\title{Reads data from google drive for the cloud viewer}
\usage{
drive_read_trap(obs)
}
\arguments{
\item{obs}{}
}
\value{

}
\description{
Reads data from google drive for the cloud viewer
}
