% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calibrations.R
\name{equipartition}
\alias{equipartition}
\title{Equipartition method for calibrating trap stiffness}
\usage{
equipartition(data)
}
\arguments{
\item{data}{A vector of trap data}
}
\description{
Equipartition method for calibrating trap stiffness
}
\examples{
#Read in trap data something like:
#data <- readr::read_tsv('Equi.txt', col_names = c('bead', 'trap'))
todays_equi_cal <- equipartition(data$bead)

}
