% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{fit_attachment_durations}
\alias{fit_attachment_durations}
\title{Define a function to use mle to estimate detachment rate and boostrap CI}
\usage{
fit_attachment_durations(data, colorz)
}
\arguments{
\item{data}{a "measured-events.csv" dataframe. First, read in data with rbind_measured_events. Needs a time_on_s columns. You might need to add this yourself, some measured events files have "time on" in miliseconds.}

\item{colorz}{a vector of colors for plotting. Should be same length as number of conditions}
}
\description{
Define a function to use mle to estimate detachment rate and boostrap CI
}
