% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_fit-bell.R
\name{fit_force_clamp}
\alias{fit_force_clamp}
\title{Fit bell equation with mle and bootstrap CI}
\usage{
fit_force_clamp(
  measured_events,
  tmin = NULL,
  plot_colors = NULL,
  basesize = 12,
  textsize = 10,
  is_shiny = FALSE
)
}
\arguments{
\item{tmin}{numeric specifying minimum attachment time (dead time)}

\item{plot_colors}{a character vector specyfing plot colors}

\item{basesize}{numeric specifying base size for plot}

\item{textsize}{numeric specifying text size for plot}

\item{force}{numeric vector containing the x-axis force measurements. Must be same length as time and id.}

\item{time}{numeric vector containing the y-axis time measurements. Must be same lenfth as force and id.}

\item{id}{NULL (default), or a character vector denoting which group the data is in. Must be same length as force and time.}
}
\value{
a list containing "plot" and "boot_data". "plot" - is a ggplot2 object. "boot_data" - is a nested data frame containing information about the fits and bootstrapping.
}
\description{
Fit bell equation with mle and bootstrap CI
}
