% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hm-cp-analysis.R
\name{fit_hm_model}
\alias{fit_hm_model}
\title{Convenient wrapper to fit Hidden-Markov Model to single molecule myosin data}
\usage{
fit_hm_model(
  trap_data,
  run_mean,
  run_var,
  use_channels,
  em_random_start,
  is_shiny,
  project,
  conditions,
  date,
  obs
)
}
\arguments{
\item{trap_data}{trap-data.csv}

\item{run_mean}{running mean}

\item{run_var}{running variance}

\item{em_random_start}{TRUE/FALSE}

\item{is_shiny}{TRUE/FALSE}
}
\value{
a list of two. 1) a table with running mean, variance, and hm-model identified state. 2) variance signal-to-noise.
}
\description{
Provides a convenient wrapper to setup and fit an HM-Model to single molecule myosin trapping data.
}
