% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_covar-hm-cp.R
\name{fit_hm_model_to_covar_smooth}
\alias{fit_hm_model_to_covar_smooth}
\title{Convenient wrapper to fit Hidden-Markov Model to COVARIANCE}
\usage{
fit_hm_model_to_covar_smooth(
  covar_smooth,
  em_random_start,
  is_shiny,
  project,
  conditions,
  date,
  obs
)
}
\arguments{
\item{covar_smooth}{smoothed covariance}

\item{em_random_start}{TRUE/FALSE}

\item{is_shiny}{TRUE/FALSE}
}
\value{
a list of two. 1) a table with running mean, variance, and hm-model identified state. 2) variance signal-to-noise.
}
\description{
Provides a convenient wrapper to setup and fit an HM-Model to single molecule myosin trapping data.
}
