% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{fit_normal_variable}
\alias{fit_normal_variable}
\title{Randomly sample a normal distribution for plotting ECDF fit line}
\usage{
fit_normal_variable(measured_events, var = "displacement_nm")
}
\arguments{
\item{measured_events}{a measured_events dataframe MUST HAVE A "displacement_nm" column! This is what it uses!!!!! it is hardcoded at the moment}

\item{var}{the column/variable to fit. Default "displacement_nm". Can also do substep 1 or 2 etc if present.}
}
\value{
a dataframe with random numbers generated from a normal distrubtion with mean and sd equal to that of your conditions data
}
\description{
Randomly sample a normal distribution for plotting ECDF fit line
}
