% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mini_ensemble_analyzer.R
\name{id_mini_events}
\alias{id_mini_events}
\title{Identify mini-ensemble events}
\usage{
id_mini_events(run_mean, displacement_threshold, time_threshold)
}
\arguments{
\item{run_mean}{calculated running mean}

\item{displacement_threshold}{threshold to determine events (x)}

\item{time_threshold}{threshold to determine events (y)
#'}

\item{data}{trap data}
}
\value{
df of events
}
\description{
Identify mini-ensemble events
}
