% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{kruskal_wilcox}
\alias{kruskal_wilcox}
\title{Prep signficance table using Krusial-Wallis and Wilcox}
\usage{
kruskal_wilcox(data, conditions, y.trans)
}
\arguments{
\item{data}{From measured-events.csv}

\item{conditions}{Grouping column}

\item{y.trans}{One of rstatix::add_xy_position y.trans options}
}
\value{
a list
}
\description{
Prep signficance table using Krusial-Wallis and Wilcox
}
