% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cloud-viewer.R
\name{lasertrapr_cloud}
\alias{lasertrapr_cloud}
\title{Shiny app for viewing lasertrapr data in the "cloud"}
\usage{
lasertrapr_cloud(email)
}
\arguments{
\item{email}{a character string. Email associated with your google drive account.}
}
\description{
This is a one line shiny app. Literally, all that is required is to 
make a new shiny app project and in the app.R call this single function and then push
the app into production. Requires putting data in a lasertrapr-cloud shared google drive which 
can be accomplished with lasertrapr::push_project_to_lasertrapr_cloud(). The shared drive will be 
created automatically if it does not exist.
}
\examples{
lasertrapr_cloud("something_at_gmail.com")
}
