% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{list_files}
\alias{list_files}
\title{Lists files into a tibble}
\usage{
list_files(...)
}
\arguments{
\item{...}{arguments to list.files function}
}
\value{
a 2 column tibble
}
\description{
Lists files into a tibble
}
