% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_hm_cp_analysis.R
\name{measure_hm_events}
\alias{measure_hm_events}
\title{Measured events identified by HM-Model}
\usage{
measure_hm_events(processed_data, hm_model_results, conversion, hz, nm2pn)
}
\arguments{
\item{hm_model_results}{results of fit_hm_model()}

\item{conversion}{}

\item{hz}{}

\item{nm2pn}{}
}
\description{
Takes the results from fit_hm_model() and estiamtes displacements, time on, time off, etc.
}
