% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_mini-ensemble-analyzer.R
\name{mini_ensemble_analyzer}
\alias{mini_ensemble_analyzer}
\title{Mini Ensemble Analyzer}
\usage{
mini_ensemble_analyzer(
  opt,
  w_width_ms = 10,
  displacement_threshold = 8,
  time_threshold_ms = 10,
  f
)
}
\arguments{
\item{opt}{an options.csv file}

\item{w_width_ms}{window width in ms}

\item{displacement_threshold}{minimum displacement threshold}

\item{time_threshold_ms}{minimum time on threshold}

\item{f}{f reactive from shiny app}
}
\description{
Mini Ensemble Analyzer
}
