% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hm_model.R
\name{mod_hm_model_server}
\alias{mod_hm_model_server}
\title{hm_model Server Function}
\usage{
mod_hm_model_server(input, output, session, f)
}
\arguments{
\item{input, output, session, f}{module parameters}
}
\description{
hm_model Server Function
}
