% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mini_ensemble.R
\name{mod_mini_ensemble_server}
\alias{mod_mini_ensemble_server}
\title{mini_ensemble Server Function}
\usage{
mod_mini_ensemble_server(input, output, session, f)
}
\arguments{
\item{input, output, session, f}{module parameters}
}
\description{
mini_ensemble Server Function
}
