% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean.R
\name{move_obs}
\alias{move_obs}
\title{Create new observation with selected data}
\usage{
move_obs(
  f,
  trap_selected_date,
  trap_obs,
  trap_selected_obs,
  trap_files,
  dygraph_clean_date_window_1,
  dygraph_clean_date_window_2
)
}
\arguments{
\item{f}{the reactiveValues list 'f' with all user selected folder info}

\item{trap_selected_date}{path to current date folder}

\item{trap_obs}{all trap observations}

\item{trap_selected_obs}{selected observation}

\item{trap_files}{all raw Data...csv files in selected obsrevation}

\item{dygraph_clean_date_window_1}{beginning index to move}

\item{dygraph_clean_date_window_2}{end index to move}
}
\description{
Create new observation with selected data
}
