% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{plot_ecdf}
\alias{plot_ecdf}
\title{Quick ECDF plot}
\usage{
plot_ecdf(measured_events, var, colorz, x_lab = "x_label", title, basesize)
}
\arguments{
\item{measured_events}{a measured events dataframe}

\item{var}{a column from measured events to plot (e.g. "displacement_nm")}

\item{colorz}{vector of colors for plot}

\item{x_lab}{a character string for x-axis label}

\item{title}{a character string for title of plot}

\item{basesize}{the size of the text}
}
\value{
a ggplot2 object
}
\description{
Quick ECDF plot
}
