% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_overlay}
\alias{plot_overlay}
\title{Plot trap data with ID'd events}
\usage{
plot_overlay(obs_path, time_period_dp, color, bead_offset = 0)
}
\arguments{
\item{obs_path}{character string of a file path to a lasertrapr observation folder}

\item{time_period_dp}{numeric vector of length 2 which denotes start/end of the trace to plot}

\item{color}{character string of a valid hex color}
}
\value{
a ggplot object
}
\description{
Plot trap data with ID'd events
}
