% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_cloud-viewer.R
\name{push_project_to_lasertrapr_cloud}
\alias{push_project_to_lasertrapr_cloud}
\title{Upload a project to be used with lasertrapr cloud}
\usage{
push_project_to_lasertrapr_cloud(project, email)
}
\arguments{
\item{project}{a character string. a lasertrapr project folder.}

\item{email}{a character string. email account associated with your google drive.}
}
\value{
nothing
}
\description{
Originally created for use with the lasertrapr cloud viewer, but can be used to backup data to goolge drive as well. 
If local changes are made and you want to update the google drive archive, running this function again will trash the old project folder on googledrive
and recreate it from scratch. Data is placed in a shared drive called "lasertrapr-cloud" on googledrive. If this folder does not exist it will be created.
}
\examples{
push_project_to_lasertrapr_cloud("project_something", "myemail")
}
