% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{rbind_measured_events}
\alias{rbind_measured_events}
\title{Combine and save all "measured-events.csv" files}
\usage{
rbind_measured_events(project, save_to_summary = FALSE)
}
\arguments{
\item{project}{a lasertrapr project folder}

\item{save_to_summary}{a logical. If TRUE will save to project/summary folder}
}
\value{
df or nothing. saves a .csv to the project/summary folder named "{date}_{project}_all-measured-events.csv"
}
\description{
Combine and save all "measured-events.csv" files
}
\examples{
rbind_measured_events(project = "project_my-project")
}
