% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simple_upload.R
\name{simple_upload}
\alias{simple_upload}
\title{Simple Upload of laser trap data}
\usage{
simple_upload(
  input_data,
  project,
  conditions,
  date,
  ready_for_analysis,
  nm2pn,
  hz
)
}
\arguments{
\item{input_data}{uploaded data to app from fileInput}

\item{project}{user selected project folder}

\item{conditions}{user selected conditions}

\item{date}{user selected date}
}
\value{
nothing. saves data to lasertrapr folder
}
\description{
Simple Upload of laser trap data
}
