% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulate.R
\name{simulate_single_molecule_trap_data}
\alias{simulate_single_molecule_trap_data}
\title{Simulate Single Molecule Trap Data}
\usage{
simulate_single_molecule_trap_data(
  n,
  hz,
  signal_to_noise,
  baseline,
  displacement,
  pi_release,
  adp_release,
  atp_binding,
  time_off
)
}
\arguments{
\item{n}{an integer. The number of events.}

\item{hz}{an integer. Sampling frequency to simulate in Hz.}

\item{signal_to_noise}{an integer. Ratio of the signal to noise.}

\item{baseline}{a list. Containing mean and sd describing the baseline noise.}

\item{displacement}{a list. Containing mean and sd describing the displacement distribution (gaussian).}

\item{pi_release}{a list. Containing the rate, lower, and upper limits (in milliseconds) to the time on distribution (exponential).}

\item{adp_release}{a list. Containing the rate, lower, and upper limits (in milliseconds) to the time off distribution (exponential).}
}
\value{
a list containing a vector of simulated data and plots of the distributions from 10000 random draws.
}
\description{
Simulate Single Molecule Trap Data
}
\examples{
simulate_single_molecule_trap_data(n = 100, 
                                             signal_to_noise = 2.5,
                                             hz = 5000,
                                             baseline = list(mean = 0, sd = 8), 
                                             time_off = list(rate = 50, lower = 15, upper = 5000),
                                             displacement = list(mean = 6, sd = 8), 
                                             pi_release = list(rate = 200, lower = 0, upper = 1500, occurs = 'after'),
                                             adp_release = list(rate = 20, lower = 0, upper = 1000, hitch = 2),
                                             atp_binding = list(rate = 1000, lower = 0, upper = 10000, hitch = 2) )

}
