% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{split_conditions_column}
\alias{split_conditions_column}
\title{Quick Summary}
\usage{
split_conditions_column(measured_events, var_names, sep)
}
\arguments{
\item{all_measured_events}{a df made from rbind_measured_events}
}
\description{
Quick Summary
}
