% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_split_obs.R
\name{split_obs}
\alias{split_obs}
\title{Split trap data into observations}
\usage{
split_obs(input_data, project, conditions, date, threshold)
}
\arguments{
\item{input_data}{the shiny path to the user uploaded data}

\item{date}{a file path to the date folder where data should be moved}

\item{threshold}{a numeric value specifying number of seconds to divide observations by}
}
\description{
Split trap data into observations
}
