% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{stats_plot_event_frequency}
\alias{stats_plot_event_frequency}
\title{Title}
\usage{
stats_plot_event_frequency(event_file_paths, factor_order, plot_colors)
}
\arguments{
\item{event_file_paths}{}

\item{factor_order}{}

\item{plot_colors}{}
}
\value{

}
\description{
Title
}
