% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{stats_plot_time_off}
\alias{stats_plot_time_off}
\title{Title}
\usage{
stats_plot_time_off(event_files_filtered, plot_colors)
}
\arguments{
\item{event_files_filtered}{}

\item{plot_colors}{}
}
\value{

}
\description{
Title
}
