% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{stats_plot_time_on}
\alias{stats_plot_time_on}
\title{Title}
\usage{
stats_plot_time_on(event_files_filtered, plot_colors, p_adj)
}
\arguments{
\item{event_files_filtered}{}

\item{plot_colors}{}

\item{p_adj}{}
}
\value{

}
\description{
Title
}
