% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{str_trap}
\alias{str_trap}
\title{Extract date time from trap data files}
\usage{
str_trap(x)
}
\arguments{
\item{x}{a laser trap raw data file name}
}
\description{
Extract date time from trap data files
}
