% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{summarize_trap_data}
\alias{summarize_trap_data}
\title{Summarize trap project}
\usage{
summarize_trap_data(f, hz, factor_order, is_shiny = T)
}
\arguments{
\item{f}{reactive folder list}

\item{hz}{sampling frequency}

\item{factor_order}{user supplied order of factors}
}
\description{
Summarize trap project
}
