% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{time_off_ecdf}
\alias{time_off_ecdf}
\title{Time Off ECDF}
\usage{
time_off_ecdf(event_files_filtered, plot_colors)
}
\arguments{
\item{event_files_filtered}{event data}

\item{plot_colors}{user colors}
}
\value{
a ggplot
}
\description{
Time Off ECDF
}
