% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{total_trap_time}
\alias{total_trap_time}
\title{Calculate Total Time of Data Collection for each condition}
\usage{
total_trap_time(project, is_shiny = FALSE)
}
\arguments{
\item{project}{character. a lasertrapr project name}

\item{is_shiny}{logical}
}
\description{
Calculate Total Time of Data Collection for each condition
}
