% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean.R
\name{trim_obs}
\alias{trim_obs}
\title{Delete selected data from observation}
\usage{
trim_obs(trap_selected_obs, trim_from, trim_to, f, hz)
}
\arguments{
\item{trap_selected_obs}{current selected observation}

\item{trim_from}{begin index to delete}

\item{trim_to}{end index to delete}

\item{f}{the reactiveValues list 'f' with all user selected folder info}
}
\description{
Delete selected data from observation
}
