% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_update-to-options.R
\name{update_to_options}
\alias{update_to_options}
\title{Update files to new options format}
\usage{
update_to_options(project_path)
}
\arguments{
\item{project_path}{a character string giving path to a lasertrapr project folder}
}
\description{
Update files to new options format
}
