% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simple-upload.R
\name{upload_data_cal_in_header}
\alias{upload_data_cal_in_header}
\title{Read in data with calibration info in the header}
\usage{
upload_data_cal_in_header(
  input_data,
  h = h,
  project,
  conditions,
  date,
  number_of_channels,
  downsample_by
)
}
\arguments{
\item{input_data}{a dataframe returned by shinyFiles}

\item{h}{user inputted header lines list. values are line/row numbers}

\item{project}{}

\item{conditions}{}

\item{date}{}

\item{number_of_channels}{}
}
\value{
nothing. saves data to lasertrapr folder
}
\description{
Read in data with calibration info in the header
}
