% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ensemble_average.R
\name{avg_ensembles}
\alias{avg_ensembles}
\title{Title}
\usage{
avg_ensembles(f, is_shiny = TRUE)
}
\arguments{
\item{is_shiny}{}

\item{trap_selected_project}{}

\item{hz}{}
}
\value{

}
\description{
Title
}
