% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calibrations.R
\name{equipartition}
\alias{equipartition}
\title{Equipartition method for calibrating trap stiffness}
\usage{
equipartition(data)
}
\arguments{
\item{data}{a vector of trap data}
}
\description{
Equipartition method for calibrating trap stiffness
}
