% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clean.R
\name{move_obs}
\alias{move_obs}
\title{Move selected data to new obs}
\usage{
move_obs(
  trap_selected_date,
  trap_selected_obs,
  trim_from,
  trim_to,
  f,
  trap_obs,
  hz
)
}
\arguments{
\item{trap_selected_date}{current selected date folder}

\item{trap_selected_obs}{current selected observation}

\item{trim_from}{begin index to delete}

\item{trim_to}{end index to delete}

\item{f}{the reactiveValues list 'f' with all user selected folder info}

\item{trap_obs}{obs folder numbers}
}
\description{
Move selected data to new obs
}
