% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{plot_overlay}
\alias{plot_overlay}
\title{Plot trap data with ID'd events}
\usage{
plot_overlay(obs_path, time_period_dp, color)
}
\arguments{
\item{obs_path}{}

\item{time_period_dp}{}

\item{color}{}

\item{axes}{}
}
\description{
Plot trap data with ID'd events
}
