% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{split_conditions_column}
\alias{split_conditions_column}
\title{Quick Summary}
\usage{
split_conditions_column(df, var_names, sep)
}
\arguments{
\item{df}{a df with conditions column to split}
}
\description{
Quick Summary
}
