% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calibrations.R
\name{step_cal}
\alias{step_cal}
\title{Performs Step Calibration of the Laser Trap}
\usage{
step_cal(vector, step)
}
\arguments{
\item{vector}{a vector of trap data}
}
\description{
Performs Step Calibration of the Laser Trap
}
