% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarize.R
\name{summarize_trap}
\alias{summarize_trap}
\title{Quick Summary}
\usage{
summarize_trap(all_measured_events, by)
}
\arguments{
\item{all_measured_events}{a df made from rbind_measured_events}
}
\description{
Quick Summary
}
